% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Title}
\usage{
clean(x, burnin = 0, subsample = 1)
}
\arguments{
\item{x}{The output of \code{simulate_posterior()}}

\item{burnin}{The number of observations to disgard at the beginning of the
chain.}

\item{subsample}{The number of observations to skip to reduce serial correlation.}
}
\value{
A cleaned up chain of draws from the posterior distribution.
}
\description{
Title
}
