% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{define}
\alias{define}
\title{Define a likelihood}
\usage{
define(x, likelihood)
}
\arguments{
\item{x}{A \code{date.frame}.}

\item{likelihood}{A formula where the right hand side is a variable in
\code{x} and the left hand side is a the name of a function.}
}
\value{
A \code{data.frame} with the likelihood as an attribute.
}
\description{
Define a likelihood
}
