% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_posterior.R
\name{draw}
\alias{draw}
\title{Simulate posterior}
\usage{
draw(x, ...)
}
\arguments{
\item{x}{The output of \code{assume()}}

\item{...}{Parameters to be passed to \code{mcmc::metrop()}.}
}
\value{
A dataframe of draws from the posterior distribution.
}
\description{
Simulate the posterior distribution
}
