% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxis}
\alias{taxis}
\title{Taxi Data}
\format{A data frame}
\usage{
data("taxis")
}
\description{
This data frame contains information about yellow taxi rides given over the
course of approximately an hour in New York City.
}
\keyword{datasets}
