% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assume.R
\name{assume}
\alias{assume}
\title{Assume}
\usage{
assume(x, prior)
}
\arguments{
\item{x}{The output of \code{\link{define}()}.}

\item{prior}{A formula with the left hand side as a \code{purrr}-style
anonymous function using the variable \code{.theta} that expresses the
prior using functions from R. For example: \code{~dnorm(.theta)}.}
}
\value{
The dataframe \code{x} with the prior as an attribute.
}
\description{
Assume a posterior distribution.
}
