% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{define}
\alias{define}
\title{Define}
\usage{
define(x, likelihood)
}
\arguments{
\item{x}{A \code{data.frame}.}

\item{likelihood}{A formula where the right hand side is a variable in
\code{x} and the left hand side is a the name of a function.}
}
\value{
A \code{data.frame} with the likelihood as an attribute.
}
\description{
Define a likelihood
}
\examples{
  my_lik <- function(data, theta) {dnorm(data, mean = 3.07333 , sd = theta)}
  my_prior <- function(theta) {dgamma(theta, shape = 1)}

  posterior <- define(iris, Sepal.Width ~ my_lik) \%>\%
    assume(prior = ~ my_prior) \%>\%
    simulate_posterior(initial = .43, nbatch = 1e5, blen = 1, scale = .01) \%>\%
    diagnose() \%>\%
    clean(burnin = 0, subsample = 40) \%>\%
    diagnose()
}
